FLINK-5976 [tests] Deduplicate Tokenizer in tests
FLINK-5976 [tests] Deduplicate Tokenizer in tests

FLINK-5976 [tests] Deduplicate Tokenizer in tests

FLINK-5976 [tests] Deduplicate Tokenizer in tests

FLINK-5976 [tests] Deduplicate Tokenizer in tests-move to testfunctions
